requirements.txt
requests==2.26.0
anthropic==0.3.0
pydantic<2.0.0
beautifulsoup4==4.10.0
openai
pydub
python-dotenv
unrealspeech
playsound 
simpleaudio
PyObjC